Correção do ajuste de horário
Programa está funcionando sem a criptografia

+++ src/Sockets/Controller/RelogioVirtual.java
@@ -9,13 +9,13 @@ import java.util.TimerTask;
 public class RelogioVirtual extends TimerTask {
 
     //variável que armazenará o tempo do relógio digital
-    private int segundos;
+    private Long segundos;
 
     public RelogioVirtual(Boolean aleatoriedade) {
 
         //Aqui iniciamos os segundos do relógio com o tempo atual do sistema
         Date dataSistema = new Date(System.currentTimeMillis());
-        this.segundos = dataSistema.getHours() * 3600 + dataSistema.getMinutes() * 60 + dataSistema.getSeconds();
+        this.segundos = dataSistema.getHours() * 3600L + dataSistema.getMinutes() * 60L + dataSistema.getSeconds();
 
         //caso o parâmetro passado pelo construtor o rellógio digital irá iniciar com um valor aleatório entre +60 e -60 segundo
         if (aleatoriedade) {
@@ -27,18 +27,18 @@ public class RelogioVirtual extends TimerTask {
     }
 
     //essa função retorna as horas
-    public int getHoras() {
-        return this.segundos / 3600;
+    public Long getHoras() {
+        return this.segundos / 3600L;
     }
 
     //essa função retorna os minutos
-    public int getMinutos() {
-        return (this.segundos - (this.getHoras() * 3600)) / 60;
+    public Long getMinutos() {
+        return (this.segundos - (this.getHoras() * 3600L)) / 60L;
     }
 
     //essa função retorna os segundos
-    public int getSegundos() {
-        return this.segundos - (this.getHoras() * 3600) - (this.getMinutos() * 60);
+    public Long getSegundos() {
+        return this.segundos - (this.getHoras() * 3600L) - (this.getMinutos() * 60L);
     }
 
     //essa função irá retornar o tempo no formato hora:minuto:segundo
@@ -46,22 +46,22 @@ public class RelogioVirtual extends TimerTask {
         return "" + this.getHoras() + ":" + this.getMinutos() + ":" + this.getSegundos();
     }
 
-    public int getTempo(){
+    public Long getTempo(){
         return this.segundos;
     }
 
 
     //função realizar soma na variável de tempo, serve para realizar ajustes e incrementar o tempo
-    public synchronized void somarTempo(int segundos) {
+    public synchronized void somarTempo(Long segundos) {
         this.segundos += segundos;
-        if (this.segundos >= 86400)
-            this.segundos += -86400 * (this.segundos/86400);
+        if (this.segundos >= 86400L)
+            this.segundos += -86400L * (this.segundos/86400L);
 
     }
 
     @Override
     public void run() {
         //essa chamada de função é feita a cada um segundo
-        somarTempo(1);
+        somarTempo(1L);
     }
 }
